Rename TextInputMolecule into LabelledTextInput
