+++ src/components/aeLabelledTextInput/index.js
@@ -1,4 +1,4 @@
-import AeTextInputMolecule from './aeTextInputMolecule.vue'
+import AeTextInputMolecule from './aeLabelledTextInput.vue'
 
 export default function install (Vue) {
   Vue.component('ae-text-input-module', AeTextInputMolecule)
